le relazioni n to n

+++ pizzeria/src/main/java/it/pizzeria/pizzeria/repository/PizzaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.pizzeria.pizzeria.model.Pizza;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Integer>{
+public interface PizzaRepository extends JpaRepository<Pizza, Long>{
     
     public List<Pizza> findByNomeContainingIgnoreCase(String nome);
